implemented confirmation to delte accounts

+++ components/data-table.tsx
@@ -25,6 +25,7 @@ import { Button } from "./ui/button";
 import { Input } from "@/components/ui/input";
 import { useState } from "react";
 import { Trash } from "lucide-react";
+import { useConfirm } from "@/hooks/use-confirm";
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
@@ -41,6 +42,10 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmDialog, confirm] = useConfirm(
+    "Are you sure you want to delete the selected rows?",
+    "You are about to delete the selected rows. This action cannot be undone."
+  );
   const [sorting, setSorting] = useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
   const [rowSelection, setRowSelection] = useState({});
@@ -60,6 +65,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
       <div className="flex items-center gap-2">
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -70,19 +76,23 @@ export function DataTable<TData, TValue>({
           className="max-w-sm mb-3"
         />
         {table.getFilteredSelectedRowModel().rows.length > 0 && (
-            <Button
-             disabled={disabled}
-             size="sm"
-             variant="outline"
-             className="ml-auto font-normal text-xs"
-             onClick={() => {
-               onDelete(table.getFilteredSelectedRowModel().rows)
-               table.resetRowSelection();
-             }}
-            >
-              Delete ({table.getFilteredSelectedRowModel().rows.length})
-              <Trash className="size-4 mr-2"/>
-            </Button>
+          <Button
+            disabled={disabled}
+            size="sm"
+            variant="outline"
+            className="ml-auto font-normal text-xs"
+            onClick={async () => {
+              const ok = await confirm();
+
+              if (ok) {
+                onDelete(table.getFilteredSelectedRowModel().rows);
+                table.resetRowSelection();
+              }
+            }}
+          >
+            Delete ({table.getFilteredSelectedRowModel().rows.length})
+            <Trash className="size-4 mr-2" />
+          </Button>
         )}
       </div>
       <div className="rounded-md border">
